requirements.txt
tensorflow==2.15.0
streamlit==1.32.2
numpy
pandas
pillow
